requirements.txt
streamlit
pandas
plotly
stripe
anthropic
psycopg2-binary
sqlalchemy
gunicorn
fastapi
uvicorn
python-dotenv
yfinance